kaixo/requirements.txt
fastapi
uvicorn[standard]
PyJWT
mangum
aws-lambda-powertools
auth0-python